requirements.txt
amadeus==11.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
filelock==3.17.0
Flask==3.1.0
fonttools==4.54.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mysql-connector-python==9.2.0
numpy==2.2.3
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scipy==1.15.2
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.1
Werkzeug==3.1.3
